exp85: fixed keyboard problem, added save-state support.
